# src/main/webapp/app/shared/api/generated/API-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "OncoKB™ is a precision oncology knowledge base developed at Memorial Sloan Kettering Cancer Center that contains biological and clinical information about genomic alterations in cancer.",
    "version": "v1.0.0",
    "title": "OncoKB™ APIs",
    "contact": {
      "name": "OncoKB™",
      "url": "https://www.oncokb.org",
      "email": "contact@oncokb.org"
    },
    "license": {
      "name": "Terms of Use",
      "url": "https://www.oncokb.org/terms"
    }
  },
...
  "paths": {
...
    "/api/account/active-trial/init": {
      "post": {
        "tags": [
          "account-resource"
        ],
        "summary": "initiateTrialAccountActivation",
        "operationId": "initiateTrialAccountActivationUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "login",
            "description": "login",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/tokens/{uuid}": {
      "get": {
        "tags": [
          "token-resource"
        ],
        "summary": "getToken",
        "operationId": "getTokenUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "uuid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Token"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/users/authorities": {
      "get": {
        "tags": [
          "user-resource"
        ],
        "summary": "getAuthorities",
        "operationId": "getAuthoritiesUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/users/{login}": {
      "get": {
        "tags": [
          "user-resource"
        ],
        "summary": "getUser",
        "operationId": "getUserUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "description": "login",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "user-resource"
        ],
        "summary": "deleteUser",
        "operationId": "deleteUserUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "description": "login",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
    "Activation": {
      "type": "object",
      "properties": {
        "activationDate": {
          "type": "string",
          "format": "date-time"
        },
        "initiatedBy": {
          "type": "string"
        },
        "initiationDate": {
          "type": "string",
          "format": "date-time"
        },
        "key": {
          "type": "string"
        }
      },
      "title": "Activation"
    },
    "AdditionalInfoDTO": {
      "type": "object",
      "properties": {
        "apiAccessRequest": {
          "$ref": "#/definitions/ApiAccessRequest"
        },
        "sentToRocReview": {
          "type": "boolean"
        },
        "trialAccount": {
          "$ref": "#/definitions/TrialAccount"
        },
        "userCompany": {
          "$ref": "#/definitions/UserCompany"
        }
      },
      "title": "AdditionalInfoDTO"
    },
    "ApiAccessRequest": {
      "type": "object",
      "properties": {
        "justification": {
          "type": "string"
        },
        "requested": {
          "type": "boolean"
        }
      },
      "title": "ApiAccessRequest"
    },
    "CompanyDTO": {
      "type": "object",
      "required": [
        "companyType",
        "licenseModel",
        "licenseStatus",
        "licenseType",
        "name"
      ],
      "properties": {
        "businessContact": {
          "type": "string"
        },
        "companyDomains": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "companyType": {
          "type": "string",
          "enum": [
            "PARENT",
            "BRANCH",
            "UNKNOWN"
          ]
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "legalContact": {
          "type": "string"
        },
        "licenseModel": {
          "type": "string",
          "enum": [
            "FULL",
            "LIMITED"
          ]
        },
        "licenseStatus": {
          "type": "string",
          "enum": [
            "TRIAL",
            "REGULAR",
            "TRIAL_EXPIRED",
            "EXPIRED",
            "UNKNOWN"
          ]
        },
        "licenseType": {
          "type": "string",
          "enum": [
            "ACADEMIC",
            "COMMERCIAL",
            "RESEARCH_IN_COMMERCIAL",
            "HOSPITAL"
          ]
        },
        "name": {
          "type": "string"
        },
        "numberOfUsers": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "CompanyDTO"
    },
...
    "Contact": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      },
      "title": "Contact"
    },
...
    "LicenseAgreement": {
      "type": "object",
      "properties": {
        "acceptanceDate": {
          "type": "string",
          "format": "date-time"
        },
        "name": {
          "type": "string"
        },
        "version": {
          "type": "string"
        }
      },
      "title": "LicenseAgreement"
    },
...
    "Token": {
      "type": "object",
      "required": [
        "currentUsage",
        "renewable"
      ],
      "properties": {
        "creation": {
          "type": "string",
          "format": "date-time"
        },
        "currentUsage": {
          "type": "integer",
          "format": "int32"
        },
        "expiration": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "renewable": {
          "type": "boolean"
        },
        "token": {
          "type": "string",
          "format": "uuid"
        },
        "usageLimit": {
          "type": "integer",
          "format": "int32"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "title": "Token"
    },
    "TrialAccount": {
      "type": "object",
      "properties": {
        "activation": {
          "$ref": "#/definitions/Activation"
        },
        "licenseAgreement": {
          "$ref": "#/definitions/LicenseAgreement"
        }
      },
      "title": "TrialAccount"
    },
...
    "User": {
      "type": "object",
      "required": [
        "activated",
        "login"
      ],
      "properties": {
        "activated": {
          "type": "boolean"
        },
        "email": {
          "type": "string",
          "minLength": 5,
          "maxLength": 254
        },
        "firstName": {
          "type": "string",
          "minLength": 0,
          "maxLength": 50
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "imageUrl": {
          "type": "string",
          "minLength": 0,
          "maxLength": 256
        },
        "langKey": {
          "type": "string",
          "minLength": 2,
          "maxLength": 10
        },
        "lastName": {
          "type": "string",
          "minLength": 0,
          "maxLength": 50
        },
        "login": {
          "type": "string",
          "minLength": 1,
          "maxLength": 50,
          "pattern": "^[_.\\-\\+@A-Za-z0-9]*$"
        },
        "resetDate": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "User"
    },
    "UserCompany": {
      "type": "object",
      "properties": {
        "anticipatedReports": {
          "type": "string"
        },
        "businessContact": {
          "$ref": "#/definitions/Contact"
        },
        "description": {
          "type": "string"
        },
        "size": {
          "type": "string"
        },
        "useCase": {
          "type": "string"
        }
      },
      "title": "UserCompany"
    },
    "UserDTO": {
      "type": "object",
      "properties": {
        "activated": {
          "type": "boolean"
        },
        "activationKey": {
          "type": "string"
        },
        "additionalInfo": {
          "$ref": "#/definitions/AdditionalInfoDTO"
        },
        "authorities": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "city": {
          "type": "string",
          "minLength": 2,
          "maxLength": 2147483647
        },
        "company": {
          "$ref": "#/definitions/CompanyDTO"
        },
        "companyName": {
          "type": "string",
          "minLength": 2,
          "maxLength": 2147483647
        },
        "country": {
          "type": "string",
          "minLength": 2,
          "maxLength": 2147483647
        },
        "createdBy": {
          "type": "string"
        },
        "createdDate": {
          "type": "string",
          "format": "date-time"
        },
        "email": {
          "type": "string",
          "minLength": 5,
          "maxLength": 254
        },
        "emailVerified": {
          "type": "boolean"
        },
        "firstName": {
          "type": "string",
          "minLength": 2,
          "maxLength": 50
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "imageUrl": {
          "type": "string",
          "minLength": 0,
          "maxLength": 256
        },
        "jobTitle": {
          "type": "string"
        },
        "langKey": {
          "type": "string",
          "minLength": 2,
          "maxLength": 10
        },
        "lastModifiedBy": {
          "type": "string"
        },
        "lastModifiedDate": {
          "type": "string",
          "format": "date-time"
        },
        "lastName": {
          "type": "string",
          "minLength": 2,
          "maxLength": 50
        },
        "licenseType": {
          "type": "string",
          "enum": [
            "ACADEMIC",
            "COMMERCIAL",
            "RESEARCH_IN_COMMERCIAL",
            "HOSPITAL"
          ]
        },
        "login": {
          "type": "string",
          "minLength": 1,
          "maxLength": 50,
          "pattern": "^[_.\\-\\+@A-Za-z0-9]*$"
        },
        "resetDate": {
          "type": "string",
          "format": "date-time"
        },
        "resetKey": {
          "type": "string"
        }
      },
      "title": "UserDTO"
    },
...
  }
}